dev-python/transitions: Add a hack removal check

+++ dev-python/transitions/transitions-0.9.1.ebuild
@@ -40,6 +40,7 @@ src_prepare() {
 	distutils-r1_src_prepare
 
 	# magic to ignore async tests sometimes, breaks EPYTEST_IGNORE
+	[[ ${PV} != 0.9.1 ]] && die "Remove this hack"
 	rm conftest.py || die
 }
